HACK: Escape specific underscore in install guide
The specific combination of underscores on lines 2895-2897 exposes a bug in the Prettier extension for Visual Studio Code.

When formatting the Markdown, the underscore in "WSS_Logging" is replaced with an asterisk, the trailing underscore in "_not_" is also replaced with an asterisk, and the leading underscore in "_not_" is escaped:

  ...the WSS*Logging database...is \_not* recommended...

I originally suspected this might be due to the escaped asterisk at the beginning of the line, but replacing that with a completely different string (i.e. "\*" --> "&dagger;") yielded the same corruption when formatting the Markdown.